Move CodeSettings widget, renaming for consistency
