+++ aiidalab_ispg/steps.py
@@ -22,7 +22,6 @@ from aiida.orm import WorkChainNode
 from aiida.plugins import DataFactory, WorkflowFactory
 from aiidalab_widgets_base import (
     AiidaNodeViewWidget,
-    ComputationalResourcesWidget,
     ProcessMonitor,
     ProcessNodesTreeWidget,
     WizardAppWidgetStep,
@@ -30,8 +29,9 @@ from aiidalab_widgets_base import (
 
 import aiidalab_ispg.qeapp as qeapp
 from aiidalab_ispg.parameters import DEFAULT_PARAMETERS
-from aiidalab_ispg.widgets import ResourceSelectionWidget
-from aiidalab_ispg.widgets import QMSelectionWidget, ExcitedStateMethod
+from .widgets import ResourceSelectionWidget
+from .widgets import QMSelectionWidget, ExcitedStateMethod
+from .input_widgets import CodeSettings
 
 from .utils import get_formula, calc_boltzmann_weights, AUtoKJ
 
@@ -151,49 +151,6 @@ class WorkChainSettings(ipw.VBox):
             ],
             **kwargs,
         )
-
-
-class CodeSettings(ipw.VBox):
-
-    codes_title = ipw.HTML(
-        """<div style="padding-top: 0px; padding-bottom: 0px">
-        <h4>Codes</h4></div>"""
-    )
-    codes_help = ipw.HTML(
-        """<div style="line-height: 140%; padding-top: 0px;
-        padding-bottom: 10px"> Select the ORCA code.</div>"""
-    )
-
-    # In the order of priority, we will select the default ORCA code from these
-    # First, we try to use SLURM on local machine, if available
-    _DEFAULT_ORCA_CODES = ("orca@slurm", "orca@localhost")
-
-    def __init__(self, **kwargs):
-
-        self.orca = ComputationalResourcesWidget(
-            default_calc_job_plugin="orca.orca",
-            description="ORCA program",
-        )
-        self._set_default_codes()
-        super().__init__(
-            children=[
-                self.codes_title,
-                self.codes_help,
-                self.orca,
-            ],
-            **kwargs,
-        )
-
-    def _set_default_codes(self):
-        for code_label in self._DEFAULT_ORCA_CODES:
-            try:
-                self.orca.value = load_code(code_label).uuid
-                return
-            except (NotExistent, ValueError):
-                pass
-
-        if not self.orca.value:
-            print("WARNING: ORCA code has not been found locally")
 
 
 class SubmitWorkChainStepBase(ipw.VBox, WizardAppWidgetStep):
